removed BORLANDC && MSDOS defines

+++ bbs/vars.h
@@ -195,16 +195,11 @@ __EXTRN__ int iia;
 
 __EXTRN__ long last_iia;
 
-//#ifndef NOT_BBS
-//#include "bbs.h"
-//#include "WSession.h"
-//#endif // NOT_BBS
-
 // from version.cpp
 extern const char *wwiv_version;
 extern const char *beta_version;
 extern const char *wwiv_date;
-extern unsigned int wwiv_num_version;
+extern unsigned short wwiv_num_version;
 
 
 // confedit
